refactor(funding-platform): satisfy quality gate for reviewer inbox
- Extract getReviewerInbox into a dedicated reviewerInboxService module so the already-oversized fundingPlatformService stops growing (695 lines, under budget)
- Drop unused type exports (InboxMilestoneDetailProps, UseInboxFeedResult, StatPillTone) — used only within their own modules
- Add the missing app/community/[communityId]/manage/error.tsx route boundary
- Name error boundaries descriptively (ManageError, ReviewerInboxError) to avoid shadowing the global Error
